Edit separate folder and Fix the slide to stop.
