+++ App.jsx
@@ -2,18 +2,18 @@ import React from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import { ToastContainer, toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
-import Home from "./src/pages/Home";
+import Home from "./src/home/Home";
 import Navbar from "./src/components/Navbar";
 import Login from "./src/pages/Login";
 import Contact from "./src/pages/Contact";
-import Cart from "./src/components/cart/Cart";
+import Cart from "./src/cart/Cart";
 import Footer from "./src/components/Footer";
 import DetailModel from "./src/components/DetailModel";
 import Arttoy from "./src/pages/Arttoy";
 import Heroes from "./src/pages/Heroes";
 import Register from "./src/pages/Register";
 import Account from "./src/pages/Account";
-import { DataProvider } from "./src/components/DataProvider";
+import { DataProvider } from "./src/context/DataProvider";
 
 const App = () => {
   return (
