+++ src/cart/CartItem.jsx
@@ -1,6 +1,5 @@
-
 import { Link } from "react-router-dom";
-import { useData } from "../DataProvider";
+import { useData } from "../context/DataProvider";
 
 const CartItem = ({
   item,
