--- src/components/Subscribe.jsx
@@ -1,68 +1,0 @@
-import React from "react";
-
-function Subscribe() {
-  return (
-    <>
-      <section id="subscribe" className="bg-[#F7F7F7] p-8">
-        <div
-          id="container-text"
-          className="flex flex-col justify-center text-center"
-        >
-          <h3 className="text-[#FFA4D5] text-2xl font-semibold">Subscribe!</h3>
-          <h1 className="text-[#B47AEA] font-bold text-3xl">News Letter</h1>
-          <p className="text-[#B47AEA] text-sm lg:text-base">
-            Subscribe to our newsletter and receive all the news and exclusive
-            deals from T-Rex!
-          </p>
-        </div>
-        <div
-          id="subscribe-email"
-          className="flex flex-col lg:flex-row lg:justify-center"
-        >
-          <input
-            type="email"
-            placeholder="E-mail"
-            className="h-10 mt-4 placeholder:pl-3 rounded-full w-full lg:w-96 lg:h-12 lg:m-4 lg:px-4 lg:py-2 border border-gray-300 lg:rounded-md lg:placeholder:pl-0 focus:ring-1 focus:outline-none"
-          />
-          <button
-            type="submit"
-            id="subscribe-btn"
-            className="h-10 mt-4 bg-[#B47AEA] rounded-full lg:h-12 lg:px-12 lg:py-2  lg:rounded-md text-white lg:hover:bg-purple-600 focus:outline-none font-semibold text-xl"
-          >
-            Subscribe
-          </button>
-        </div>
-      </section>
-
-      <section id="icon-service">
-        <div
-          id="container-service"
-          className="p-2 grid grid-cols-2 lg:flex lg:flex-row lg:justify-evenly lg:p-10"
-        >
-          <div id="service-box1" className="text-center p-8 lg:p-10">
-            <i className="text-[#B47AEA] fa-solid fa-lock text-4xl"></i>
-            <p className="">Payment</p>
-            <p>100% secured</p>
-          </div>
-          <div id="service-box2" className="text-center p-8 lg:p-10">
-            <i className="text-[#B47AEA] fa-solid fa-rocket text-4xl"></i>
-            <p>Payment</p>
-            <p>100% secured</p>
-          </div>
-          <div id="service-box3" className="text-center p-8 lg:p-10">
-            <i className="text-[#B47AEA] fa-solid fa-gift text-4xl"></i>
-            <p>Payment</p>
-            <p>100% secured</p>
-          </div>
-          <div id="service-box4" className="text-center p-8 lg:p-10">
-            <i className="text-[#B47AEA] fa-solid fa-comments text-4xl"></i>
-            <p>Payment</p>
-            <p>100% secured</p>
-          </div>
-        </div>
-      </section>
-    </>
-  );
-}
-
-export default Subscribe;
