+++ src/home/Hero.jsx
@@ -1,0 +1,27 @@
+import React, { useState, useEffect } from "react";
+
+const Hero = ({ images }) => {
+  const [currentImageIndex, setCurrentImageIndex] = useState(0);
+
+  useEffect(() => {
+    const interval = setInterval(() => {
+      setCurrentImageIndex((prevIndex) => (prevIndex + 1) % images.length);
+    }, 3000);
+
+    return () => clearInterval(interval);
+  }, [images.length]);
+
+  return (
+    <section className="hero">
+      <div className="">
+        <img
+          src={images[currentImageIndex]}
+          alt={`Slide ${currentImageIndex + 1}`}
+          className="w-full h-auto"
+        />
+      </div>
+    </section>
+  );
+};
+
+export default Hero;
